pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "wizard_ops"
version = "0.0.6"
description = "Food Nutrition Predictor Using ResNet18"
authors = [
  { name = "Yubaba", email = "s223190@dtu.dk" },
  { name = "Malificent", email = "s253081@dtu.dk" },
  { name = "Eleven", email = "s253471@dtu.dk" },
  { name = "Wanda", email = "s253033@dtu.dk" },
  { name = "Merlin", email = "alihaj@dtu.dk" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "datasets>=4.4.2",
  "fastapi==0.115.6",
  "google-api-python-client>=2.188.0",
  "h5py>=3.15.1",
  "hydra-core>=1.3.2",
  "kaggle>=1.8.3",
  "kagglehub>=0.4.0",
  "loguru>=0.7.3",
  "matplotlib>=3.10.8",
  "onnx>=1.20.1",
  "onnxruntime>=1.23.2",
  "onnxscript>=0.5.7",
  "pandas>=2.2.0",
  "pillow>=12.1.0",
  "tensorboard>=2.20.0",
  "tqdm>=4.67.1",
  "transformers>=4.57.6",
  "typer>=0.21.1",
  "wandb>=0.23.1",
  "streamlit>=1.30.0",
  "python-multipart>=0.0.6",
  "requests>=2.32.5",
  "google-cloud-run>=0.14.0",
  "google-cloud-storage>=3.8.0",
  "uvicorn>=0.34.0",
  "torch>=2.10.0",
  "torchvision>=0.25.0",
  "lightning>=2.6.0",
  "dvc[gs]>=3.66.1",
  "albumentations>=2.0.8",
]

[project.scripts]
wizard_ops = "wizard_ops.cli:main"

[dependency-groups]
dev = [
  "coverage==7.6.9",
  "invoke==2.2.0",
  "mkdocs-material==9.4.6",
  "mkdocs==1.6.1",
  "pre-commit==4.1.0",
  "ruff>=0.10",
  "mkdocstrings-python==1.12.2",
  "pytest==9.0.2",
  "dvc[gs]>=3.66.1",
]

# [tool.uv.workspace]
# members = ["src/wizard_ops/frontend", "src/wizard_ops/backend"]

[tool.ruff]
line-length = 120
lint.select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]
